수정

+++ src/main/java/com/example/booking/application/PaymentConfirmRequest.java
@@ -1,7 +1,9 @@
 package com.example.booking.application;
 
+import java.math.BigDecimal;
+
 public record PaymentConfirmRequest(
         String paymentKey,
         String orderId,
-        long amount
+        BigDecimal amount
 ) {}
